Tests

+++ relationships/views.py
@@ -8,7 +8,7 @@ from auths.models import User
 from rest_framework import permissions
 
 
-class FollowRelationshipView(APIView):
+class RelationshipViewSet(viewsets.ViewSet):
     permission_classes = (permissions.IsAuthenticated,)
 
     @staticmethod
@@ -18,7 +18,7 @@ class FollowRelationshipView(APIView):
             return Response({
                 'message': 'Bad request',
                 'errors': serializer.errors
-                }, status=status.HTTP_400_BAD_REQUEST
+            }, status=status.HTTP_400_BAD_REQUEST
             )
         data = serializer.validated_data
 
@@ -29,7 +29,7 @@ class FollowRelationshipView(APIView):
         if not check_to_user:
             return Response({'error': 'No such user exists'}, status=status.HTTP_404_NOT_FOUND)
         if from_user_id == to_user_id:
-            return Response({'error': 'SELF'}, status=status.HTTP_204_NO_CONTENT)
+            return Response({'status': 'SELF'}, status=status.HTTP_204_NO_CONTENT)
 
         current_rel = Relationships.objects.filter(from_user_id=from_user_id, to_user_id=to_user_id).first()
         rel_requested = data['status']
@@ -41,9 +41,12 @@ class FollowRelationshipView(APIView):
             relationship.save()
             return Response({'status': rel_requested}, status=status.HTTP_201_CREATED)
 
+        elif not current_rel:
+            return Response({'error': 'The requests are incorrect.'}, status=status.HTTP_400_BAD_REQUEST)
+
         elif (current_rel.status == 'FOLLOW' and rel_requested == 'UNFOLLOW') or \
-             (current_rel.status == 'BLOCK' and rel_requested == 'UNBLOCK') or \
-             (current_rel.status in ['UNFOLLOW', 'UNBLOCK'] and rel_requested in ['FOLLOW', 'BLOCK']):
+                (current_rel.status == 'BLOCK' and rel_requested == 'UNBLOCK') or \
+                (current_rel.status in ['UNFOLLOW', 'UNBLOCK'] and rel_requested in ['FOLLOW', 'BLOCK']):
 
             current_rel.status = rel_requested
             current_rel.save()
@@ -51,39 +54,25 @@ class FollowRelationshipView(APIView):
         else:
             return Response({'error': 'The requests are incorrect.'}, status=status.HTTP_400_BAD_REQUEST)
 
-
-class UpdateRelationshipView(APIView):
-    permission_classes = (permissions.IsAuthenticated,)
-
-    def get_object(self, pk):
+    @staticmethod
+    def get_object(pk):
         try:
             return Relationships.objects.get(pk=pk)
         except Relationships.DoesNotExist:
             raise Http404
 
-    def put(self, request, pk):
-        relationship = self.get_object(pk)
-        rel_requested = request.data['status']
-
-        if (relationship.status == 'FOLLOW' and rel_requested == 'UNFOLLOW') or \
-           (relationship.status == 'BLOCK' and rel_requested == 'UNBLOCK') or \
-           (relationship.status in ['UNFOLLOW', 'UNBLOCK'] and rel_requested in ['FOLLOW', 'BLOCK']):
-
-            relationship.status = rel_requested
-            relationship.save()
-
-            return Response(data=RelationshipSerializer(relationship).data, status=status.HTTP_200_OK)
+    @staticmethod
+    def delete(request, pk):
+        '''Delete to be used by admins only, as customer user should be using
+        post UNFOLLOW/UNBLOCK'''
+        user_id = request.user.id
+        user_role = request.user.user_role
+        if user_role == 'superuser' or user_role == 'admin':
+            relationship = RelationshipViewSet.get_object(pk)
+            relationship.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
         else:
-            response = 'Your update request is incorrect.'
-            return Response(data=response, status=status.HTTP_400_BAD_REQUEST)
-
-    def delete(self, request, pk):
-        relationship = self.get_object(pk)
-        relationship.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
-
-
-class RelationshipViewSet(viewsets.ViewSet):
+            return Response(status=status.HTTP_400_BAD_REQUEST)
 
     @staticmethod
     def list(request):
@@ -93,21 +82,24 @@ class RelationshipViewSet(viewsets.ViewSet):
         if user_role == 'superuser' or user_role == 'admin':
             relationships = Relationships.objects.order_by('-updated_at')
             data = RelationshipSerializer(relationships, many=True).data
-
             return Response({'data': data}, status=status.HTTP_200_OK)
+
         else:
             user_name = User.objects.get(id=user_id).username
-            relationships_to = Relationships.objects.filter(from_user_id=user_id).order_by('-updated_at')
-            relationships_from = Relationships.objects.filter(to_user_id=user_id).order_by('-updated_at')
-
+            relationships_from_user = Relationships.objects.filter(from_user_id=user_id,
+                                                                   to_user__is_active=True).order_by('-updated_at')
+            relationships_to_user = Relationships.objects.filter(to_user_id=user_id,
+                                                                 from_user__is_active=True).order_by('-updated_at')
+            rel_from = RelationshipSerializer(relationships_from_user, many=True).data
+            rel_to = RelationshipSerializer(relationships_to_user, many=True).data
             return Response({
                 'user_name': user_name,
-                'relationships_to': RelationshipSerializer(relationships_to, many=True).data,
-                'relationships_from': RelationshipSerializer(relationships_from, many=True).data
+                'relationships_from_user': rel_from,
+                'relationships_to_user': rel_to
             }, status=status.HTTP_200_OK)
 
     @staticmethod
-    def retrieve(request, pk=None):
+    def retrieve(request, pk):
         user_id = request.user.id
         user_role = request.user.user_role
